+++ Kernel/System/DynamicField/Driver/BaseText.pm
@@ -183,8 +183,16 @@ sub EditFieldRender {
         $FieldClass .= ' ServerError';
     }
 
+    my $ValueEscaped = $Param{LayoutObject}->Ascii2Html(
+        Text => $Value,
+    );
+
+    my $FieldLabelEscaped = $Param{LayoutObject}->Ascii2Html(
+        Text => $FieldLabel,
+    );
+
     my $HTMLString = <<"EOF";
-<input type="text" class="$FieldClass" id="$FieldName" name="$FieldName" title="$FieldLabel" value="$Value" />
+<input type="text" class="$FieldClass" id="$FieldName" name="$FieldName" title="$FieldLabelEscaped" value="$ValueEscaped" />
 EOF
 
     if ( $Param{Mandatory} ) {